[fix] Fixed wrong assignment in create_ai function
* Fixed an issue that n_hidden_channels was assigned to n_layers
  This was cause that DQN-training became slow
* Specified observation_space dtype as float32
* Added profile to .gitignore

+++ reversi_dqn.py
@@ -15,7 +15,8 @@ class ReversiEnv(gym.core.Env):
         self.action_space = gym.spaces.Discrete(64)
         self.observation_space = gym.spaces.Box(
             low = np.zeros(8*8*4, dtype=np.float32),
-            high = np.ones(8*8*4, dtype=np.float32) )
+            high = np.ones(8*8*4, dtype=np.float32),
+            dtype = np.float32 )
 
     def reset(self, player):
         self._player = player
